Revisar CommitSession
