+++ RetroFrog V4/.gitignore
@@ -3,7 +3,7 @@ node_modules
 /.cache
 /build
 /public/build
-.env
+
 postgres-data
 
 tailwind.css
