+++ RetroFrog V4/app/routes/home.tsx
@@ -1,6 +1,6 @@
 import { NavLink, Outlet } from '@remix-run/react';
 import { useState } from 'react';
-import Custom from '~/components/Custom';
+import Custom from '~/components/Custom2';
 import {
   CollapseArrow,
   FavGamesIcon,
