+++ RetroFrog V4/app/tailwind.css
@@ -13,7 +13,6 @@ body {
 
 
 body {
-  background-image: url('/assets/background/bg.jpg');
   background-size: cover;
   background-attachment: fixed;
   background-position: center;
